Rename defaultPlugins to corePlugins
