+++ __tests__/applyAtRule.test.js
@@ -1,10 +1,10 @@
 import postcss from 'postcss'
 import substituteClassApplyAtRules from '../src/lib/substituteClassApplyAtRules'
 import processPlugins from '../src/util/processPlugins'
-import defaultPlugins from '../src/defaultPlugins'
+import corePlugins from '../src/corePlugins'
 import defaultConfig from '../defaultConfig.stub.js'
 
-const { utilities: defaultUtilities } = processPlugins(defaultPlugins(defaultConfig), defaultConfig)
+const { utilities: defaultUtilities } = processPlugins(corePlugins(defaultConfig), defaultConfig)
 
 function run(input, config = defaultConfig, utilities = defaultUtilities) {
   return postcss([substituteClassApplyAtRules(config, utilities)]).process(input, {
@@ -205,7 +205,7 @@ test('you can apply utility classes without using the given prefix', () => {
     prefix: 'tw-',
   }
 
-  return run(input, config, processPlugins(defaultPlugins(config), config).utilities).then(
+  return run(input, config, processPlugins(corePlugins(config), config).utilities).then(
     result => {
       expect(result.css).toEqual(expected)
       expect(result.warnings().length).toBe(0)
@@ -229,7 +229,7 @@ test('you can apply utility classes without using the given prefix when using a 
     },
   }
 
-  return run(input, config, processPlugins(defaultPlugins(config), config).utilities).then(
+  return run(input, config, processPlugins(corePlugins(config), config).utilities).then(
     result => {
       expect(result.css).toEqual(expected)
       expect(result.warnings().length).toBe(0)
