+++ src/corePlugins.js
